refactor: change name validator
